% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_ts.R
\name{get_available_ts}
\alias{get_available_ts}
\title{Find the available assessment time series data}
\usage{
get_available_ts(itis = NULL, jurisdiction = NULL)
}
\arguments{
\item{itis}{Numeric vector. Species ITIS code (Default = NULL, all species)}

\item{jurisdiction}{Character string. Management council}
}
\value{
data frame (n x 5)
\item{StockName}{Full name of stock}
\item{Jurisdiction}{Management council}
\item{ITIS}{species itis code}
\item{AssessmentYear}{Year of assessment}
\item{nYrs}{Average number of years data (of Abundance, Catch, Recruitment,
Fmort). Some assessments have differeing lengths of time series.  }
}
\description{
Filters the Summary data to find all time series data
}
