% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_full_assessment.r
\name{get_latest_full_assessment}
\alias{get_latest_full_assessment}
\title{Pulls the most recent full stock assessment data.}
\usage{
get_latest_full_assessment(itis = NULL)
}
\arguments{
\item{itis}{Numeric vector. Species ITIS code (Default = NULL, all species)}
}
\value{
list of 2 data frames:
\item{data}{Data frame consisting of time series values of the most recent Benchmark or Full update for the metrics Abundance, Fmort, Recruitment, and Catch.}
\item{summary}{Summary of data returned in \code{data}. StockName, CommonName, StockArea, ITIS, Metric in addition to the
AssessmentYear the data came from, the FirstYear and LastYear of the data and the number of years (numYear) of data retrieved}
}
\description{
Often a benchmark or a Full update contains the longest time series of data
whereas partial updates only contain data since last benchmark.
The most recent year in which Abundance, Catch, Fmort, Recruitment are present is returned for each species
}
