% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_metrics.r
\name{get_latest_metrics}
\alias{get_latest_metrics}
\title{Filter most recent time series data}
\usage{
get_latest_metrics(itis = NULL, metrics = c("Catch", "Abundance"))
}
\arguments{
\item{itis}{Numeric vector. Species ITIS code (Default = NULL, all species)}

\item{metrics}{Character vector. The metrics for which data are required ("Catch","Abundance","Fmort","Recruitment").
Default = c("Catch","Abundance"). All selected metrics must be available for a successful data pull for each species.}
}
\value{
list of 2 data frames:
\item{data}{Data frame consisting of time series values of the most recent Benchmark or Full update for the requested metrics.}
\item{summary}{Summary of data returned in \code{data}. StockName, CommonName, StockArea, ITIS, Metric in addition to the
AssessmentYear the data came from, the FirstYear and LastYear of the data and the number of years (numYear) of data retrieved}
}
\description{
Pulls the most recent data for set of metrics provided.
Often a benchmark or a Full update contains the longest time series of data
whereas partial updates only contain data since last benchmark.
Data from the most recent Full Update or Benchmark is returned.
}
