% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_points.R
\name{get_reference_points}
\alias{get_reference_points}
\title{Find reference points for a species stock}
\usage{
get_reference_points(stock_id = NULL, ref_point = "bmsy")
}
\arguments{
\item{stock_id}{Numeric. Stock id number, found using  `get_species_itis()``}

\item{ref_point}{Character string. The name of the reference point eg ("bmsy", "fmsy")}
}
\value{
data frame
\item{ref_point}{\code{ref_point} selected by user}
\item{assessment_year}{The year of the assessment}
\item{assessment_type}{Type of assessment conducted}
\item{assessment_model}{Type of assessment model used}
}
\description{
Filters the Summary data to find species reference point information
}
\examples{
# Get fmsy reference point for stock_id = 10509 (Atlantic cod - Georges Bank)
get_reference_points(stock_id = 10509, ref_point = "fmsy")

}
