% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_itis.R
\name{get_species_itis}
\alias{get_species_itis}
\title{Find the species itis}
\usage{
get_species_itis(itis = NULL, stock = NULL)
}
\arguments{
\item{itis}{Numeric vector. Species ITIS code (Default = NULL, all species)}

\item{stock}{Character vector. String in which to use to search for stock}
}
\value{
data frame (n x 3)
\item{StockName}{Full name of stock}
\item{Jurisdiction}{Management council}
\item{ITIS}{species itis code}
}
\description{
Filters the Summary data to find species stock information
}
