% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables}
\alias{get_tables}
\title{Extracts table from pdf}
\usage{
get_tables(pdfName, tableTypes)
}
\arguments{
\item{pdfName}{String. Name of pdf file}

\item{tableTypes}{Dataframe. Two columms. Column 1 = Type of data in table ("Biomass","Landings", etc.).
Column 2 = page number(s) in document where table exists}
}
\value{
The extracted table a sa tibble
}
\description{
Given a pdf and the page numbers, a table of data is extracted
}
\section{Input}{


Input: Pdf documents are assumed to be in the "reports" folder of the current working directory.
}

