% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series data for a particular species}
\usage{
plot_ts(
  itis = NULL,
  stock = NULL,
  metric = "Catch",
  facetplot = FALSE,
  printfig = TRUE
)
}
\arguments{
\item{itis}{Numeric. Species ITIS code}

\item{stock}{Character. Full name of stock (only required if more than one
stock exists for ITIS code)}

\item{metric}{Character vector. Specifying which metric to plot (Catch,
Abundance, Fmort, Recruitment, Index)}

\item{facetplot}{Boolean. How to plot results. Plot each assessment in its
own facet (TRUE) or pplot all assessments on a single plot (default = FALSE)}

\item{printfig}{Boolean. Print figure to figure window (Default = T)}
}
\value{
A list of two items is returned
\item{plot}{ggplot object}
\item{data}{dataframe used in the plotting}
}
\description{
Exploratory plots of Catch, Abundance, F, Recruitment for all assessments
for a particular species
}
\section{Units}{

The units of some of the metrics change over time. For example, catch for
one assessment may be presented in metric tons whereas other assessments
may be in thousand metric tons. This is not resolved in the plotting.
When this arises it is best to use \code{facet = TRUE} since
each facet uses its own yaxis scale
}

