% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_species_stock.R
\name{pull_species_stock}
\alias{pull_species_stock}
\title{Pull species data directly}
\usage{
pull_species_stock(itis = NULL, stock = NULL)
}
\arguments{
\item{itis}{Numeric vector. Species ITIS code (Default = NULL, all species)}

\item{stock}{Character vector. String in which to use to search for stock}
}
\value{
data frame (n x 3)
\item{StockName}{Full name of stock}
\item{Jurisdiction}{Management council}
\item{ITIS}{species itis code}
}
\description{
Pulls data defined by stock directly to obtain current data. For those who
don't want to reinstall the package to get most recent data
}
