% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stock_assessment_summary.R
\docType{data}
\name{stock_assessment_summary}
\alias{stock_assessment_summary}
\title{Reference points, contact info, assessment metadata for USA assessed stocks}
\format{
A data frame

\describe{
\item{stock_name}{}
\item{stock_id}{Unique Stock identifier}
\item{assessment_id}{Unique Assessment identifer}
\item{jurisdiction}{}
\item{fmp}{}
\item{science_center}{}
\item{regional_ecosystem}{}
\item{fssi_stock}{}
\item{itis}{}
\item{scientific_name}{}
\item{common_name}{}
\item{stock_area}{}
\item{assessment_year}{}
\item{assessment_month}{}
\item{last_data_year}{}
\item{assessment_type}{Joining of Update Type and Assessment Type:
Describes the level of assessment effort}
\item{review_result}{}
\item{assessment_model}{}
\item{model_version}{}
\item{lead_lab}{}
\item{citation}{}
\item{final_assessment_report_1}{}
\item{final_assessment_report_2}{}
\item{point_of_contact}{}
\item{life_history_data}{}
\item{abundance_data}{}
\item{catch_data}{}
\item{assessment_level}{}
\item{assessment_frequency}{}
\item{model_category}{}
\item{catch_input_data}{}
\item{abundance_input_data}{}
\item{biological_input_data}{}
\item{ecosystem_linkage}{}
\item{composition_input_data}{}
\item{f_year}{}
\item{estimated_f}{}
\item{f_unit}{}
\item{f_basis}{}
\item{flimit}{}
\item{flimit_basis}{}
\item{fmsy}{}
\item{fmsy_basis}{}
\item{f_over_flimit}{}
\item{f_over_fmsy}{}
\item{ftarget}{}
\item{ftarget_basis}{}
\item{f_over_ftarget}{}
\item{b_year}{}
\item{estimated_b}{}
\item{b_unit}{}
\item{b_basis}{}
\item{blimit}{}
\item{blimit_basis}{}
\item{bmsy}{}
\item{bmsy_basis}{}
\item{b_over_blimit}{}
\item{b_over_bmsy}{}
\item{msy}{}
\item{msy_unit}{}

}
}
\source{
\url{https://apps-st.fisheries.noaa.gov/stocksmart?app=homepage}
}
\usage{
stock_assessment_summary
}
\description{
Stock assessment metadata and reference points for species found in waters
of the USA. Data is provided for actively managed stocks of a federal
Fisheries Management Plan (FMP). Stocks solely managed by US States are
not provided.

For a description of all variables please refer to
the Stock SMART data dictionary
\url{https://apps-st.fisheries.noaa.gov/stocksmart/StockSMART_DataDictionary.pdf}
}
\section{Data download}{


The data were downloaded from \href{https://apps-st.fisheries.noaa.gov/stocksmart?app=browse-by-stock}{Stock SMART} - Status, Management, Assessments & Resource Trends.
}

\seealso{
Other stockAssessments:
\code{\link{stockAssessmentData}},
\code{\link{stockAssessmentSummary}},
\code{\link{stock_assessment_data}}
}
\concept{stockAssessments}
\keyword{datasets}
