% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stockAssessmentData.R
\docType{data}
\name{stockAssessmentData}
\alias{stockAssessmentData}
\title{Abundance, Catch, Fishing mortality, Recruitment for USA assessed stocks}
\format{
A data frame with n rows and m variables

\describe{
\item{StockName}{Species name and stock area for the assessed species}
\item{Stockid}{Unique stock identifier}
\item{Assessmentid}{Unique assessment identifier}
\item{CommonName}{Common name for species}
\item{ScientificName}{Scientific name for species}
\item{ITIS}{Species ITIS Taxon Serial Number }
\item{StockArea}{Stock Area: The region in which the species is assessed}
\item{Year}{Year the data was collected/generated}
\item{Value}{Value of \code{Metric}}
\item{Metric}{Typeof metric: Abundance of species, Index of Biomass,
Catch, Recruitment of juvenile individuals, Fmort - Fishing mortality}
\item{Description}{Description of \code{Metric}}
\item{Units}{Units of \code{Value}}
\item{AssessmentYear}{Year of Assessment from which the data was obtained}
\item{FMP}{Fisheries management plan species is managed under}
\item{RegionalEcosystem}{The Regional Ecosystem wher stock is found.
There are 8 ecosystems defined around the coastal margins of the
United States}
\item{Jurisdiction}{The Fishery Management Council or other group
responsible for managing this stock}
\item{AssessmentType}{Describes the level of assessment effort.
For assessments completed prior to 2018
:New, Benchmark, Full Update, Partial Update, for post 2018: Operational,
Stock Monitoring Update, Research & Operational}

}
}
\source{
\url{https://apps-st.fisheries.noaa.gov/stocksmart?app=homepage}
}
\usage{
stockAssessmentData
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{stockAssessmentData} is deprecated. It will continue to receive updates for while. From v1.0.0 it will not be used in any of the supporting functions.
Camel case naming is being replaced by snake case. Please use \link{stock_assessment_data} instead since this data set is used in all supporting functions.

Stock assessment data for species found in waters of the USA. Data is
provided for actively managed stocks of a federal Fisheries Management
Plan (FMP). Stocks solely managed by US States are not provided.
}
\section{Data download}{


The data were downloaded from \href{https://apps-st.fisheries.noaa.gov/stocksmart?app=browse-by-stock}{Stock SMART} - Status, Management, Assessments & Resource Trends.

For Manual download:
\enumerate{
\item Click the "Download Data" module
\item Select data to download (dropdown) -> Assessment Time Series Data
\item Click Search Stocks button
\item Click Select All Stocks button
\item Click Select All Asmts button
\item Click Download button
\item Download all Parts
}
}

\seealso{
Other stockAssessments:
\code{\link{stockAssessmentSummary}},
\code{\link{stock_assessment_data}},
\code{\link{stock_assessment_summary}}
}
\concept{stockAssessments}
\keyword{internal}
