% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stockAssessmentSummary.R
\docType{data}
\name{stockAssessmentSummary}
\alias{stockAssessmentSummary}
\title{Reference points, contact info, assessment metadata for USA assessed stocks}
\format{
A data frame with n rows and m variables

\describe{
\item{Stock Name}{}
\item{Stock ID}{Unique Stock identifier}
\item{Assessment ID}{Unique Assessment identifer}
\item{Jurisdiction}{}
\item{FMP}{}
\item{Science Center}{}
\item{Regional Ecosystem}{}
\item{FSSI Stock?}{}
\item{ITIS Taxon Serial Number}{}
\item{Scientific Name}{}
\item{Common Name}{}
\item{Stock Area}{}
\item{Assessment Year}{}
\item{Assessment Month}{}
\item{Last Data Year }{}
\item{Assessment Type}{Joining of Update Type and Assessment Type:
Describes the level of assessment effort}
\item{Review Result }{}
\item{Assessment Model}{}
\item{Model Version}{}
\item{Lead Lab }{}
\item{Citation}{}
\item{Final Assessment Report 1}{}
\item{Final Assessment Report 2}{}
\item{Point of Contact}{}
\item{Life History Data}{}
\item{Abundance Data}{}
\item{Catch Data}{}
\item{Assessment Level}{}
\item{Assessment Frequency}{}
\item{Model Category }{}
\item{Catch Input Data}{}
\item{Abundance Input Data}{}
\item{Biological Input Data}{}
\item{Ecosystem Linkage}{}
\item{Composition Input Data}{}
\item{F Year}{}
\item{Estimated F}{}
\item{F Unit}{}
\item{F Basis}{}
\item{Flimit}{}
\item{Flimit Basis}{}
\item{Fmsy}{}
\item{Fmsy Basis}{}
\item{F/Flimit}{}
\item{F/Fmsy}{}
\item{Ftarget}{}
\item{Ftarget Basis}{}
\item{F/Ftarget}{}
\item{B Year}{}
\item{Estimated B }{}
\item{B Unit }{}
\item{B Basis }{}
\item{Blimit }{}
\item{Blimit Basis}{}
\item{Bmsy}{}
\item{Bmsy Basis}{}
\item{B/Blimit}{}
\item{B/Bmsy}{}
\item{MSY}{}
\item{MSY Unit}{}

}
}
\source{
\url{https://apps-st.fisheries.noaa.gov/stocksmart?app=homepage}
}
\usage{
stockAssessmentSummary
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{stockAssessmentSummary} is deprecated. It will continue to receive updates for while. From v1.0.0 it will not be used in any of the supporting functions.
Camel case naming is being replaced by snake case. Please use \link{stock_assessment_summary} instead since this data set is used in all supporting functions.

Stock assessment metadata and reference points for species found in waters
of the USA. Data is provided for actively managed stocks of a federal
Fisheries Management Plan (FMP). Stocks solely managed by US States are
not provided. For a description of all variables please refer to
the Stock SMART data dictionary
\url{https://apps-st.fisheries.noaa.gov/stocksmart/StockSMART_DataDictionary.pdf}
}
\section{Data download}{


The data were downloaded from \href{https://apps-st.fisheries.noaa.gov/stocksmart?app=browse-by-stock}{Stock SMART} - Status, Management, Assessments & Resource Trends.

For Manual download:
\enumerate{
\item Click the "Download Data" module
\item Select data to download (dropdown) -> Assessment Summary Data
\item Select full range of years from Start Calendar Year- End Calendar Year
\item Click Search Stocks button
\item Click Select All Stocks button
\item Click Select All Fields button
\item Click Download button
\item Download all Parts
}
}

\seealso{
Other stockAssessments:
\code{\link{stockAssessmentData}},
\code{\link{stock_assessment_data}},
\code{\link{stock_assessment_summary}}
}
\concept{stockAssessments}
\keyword{internal}
