% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_itis.R
\name{get_species_itis}
\alias{get_species_itis}
\title{Find the species itis}
\usage{
get_species_itis(itis = NULL, stock = NULL)
}
\arguments{
\item{itis}{Numeric vector. Species itis code (Default = NULL, all species)}

\item{stock}{Vector. Either a character string or numeric in which to use to search for stock name or stock id}
}
\value{
data frame
\item{stock_name}{Full name of stock}
\item{jurisdiction}{Management council}
\item{itis}{species itis code}
\item{stock_id}{stock id code}
}
\description{
Filters the Summary data to find species stock information
}
\examples{
# Get ITIS info for Atlantic Cod, (ITIS code = 1647112)
get_species_itis(stock = "Atlantic cod")
get_species_itis(stock = "cod")
get_species_itis(itis = 167712)
# Get ITIS info for a specific cod stock
get_species_itis(stock = "12805")
get_species_itis(stock = 12805)

}
