% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_stock_data.R
\name{get_species_stock_data}
\alias{get_species_stock_data}
\title{Pull real-time species stock data directly from StockSMART API}
\usage{
get_species_stock_data(itis = NULL, stock = NULL)
}
\arguments{
\item{itis}{Numeric vector. Species ITIS code (Default = NULL, all species)}

\item{stock}{Character vector. String in which to use to search for stock}
}
\value{
data frame (n x 3)
\item{StockName}{Full name of stock}
\item{Jurisdiction}{Management council}
\item{ITIS}{species itis code}
}
\description{
Pulls data defined by stock directly to obtain current data. For those who
don't want to reinstall the package to get most recent data
}
