% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stock_assessment_data.R
\docType{data}
\name{stock_assessment_data}
\alias{stock_assessment_data}
\title{Abundance, Catch, Fishing mortality, Recruitment for USA assessed stocks}
\format{
A data frame

\describe{
\item{stock_name}{Species name and stock area for the assessed species}
\item{stock_id}{Unique stock identifier}
\item{assessment_id}{Unique assessment identifier}
\item{common_name}{Common name for species}
\item{scientific_name}{Scientific name for species}
\item{itis}{Species ITIS Taxon Serial Number }
\item{stock_area}{Stock Area: The region in which the species is assessed}
\item{year}{Year the data was collected/generated}
\item{value}{Value of \code{Metric}}
\item{metric}{Typeof metric: Abundance of species, Index of Biomass,
Catch, Recruitment of juvenile individuals, Fmort - Fishing mortality}
\item{description}{Description of \code{Metric}}
\item{units}{Units of \code{Value}}
\item{assessment_year}{Year of Assessment from which the data was obtained}
\item{fmp}{Fisheries management plan species is managed under}
\item{regional_ecosystem}{The Regional Ecosystem wher stock is found.
There are 8 ecosystems defined around the coastal margins of the
United States}
\item{jurisdiction}{The Fishery Management Council or other group
responsible for managing this stock}
\item{assessment_type}{Describes the level of assessment effort.
For assessments completed prior to 2018
:New, Benchmark, Full Update, Partial Update, for post 2018: Operational,
Stock Monitoring Update, Research & Operational}

}
}
\source{
\url{https://apps-st.fisheries.noaa.gov/stocksmart?app=homepage}
}
\usage{
stock_assessment_data
}
\description{
Stock assessment data for species found in waters of the USA. Data is
provided for actively managed stocks of a federal Fisheries Management
Plan (FMP). Stocks solely managed by US States are not provided.
}
\section{Data download}{


The data were downloaded from \href{https://apps-st.fisheries.noaa.gov/stocksmart?app=browse-by-stock}{Stock SMART} - Status, Management, Assessments & Resource Trends.
}

\seealso{
Other stockAssessments: 
\code{\link{stockAssessmentData}},
\code{\link{stockAssessmentSummary}},
\code{\link{stock_assessment_summary}}
}
\concept{stockAssessments}
\keyword{datasets}
