% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockAssessmentData.R
\docType{data}
\name{stockAssessmentData}
\alias{stockAssessmentData}
\title{Abundance, Catch, fishing mortalit, Recruitment for USA assessed stocks}
\format{
A data frame with n rows and m variables

\describe{
\item{StockName}{Species name and stock area for the assessed species}
\item{CommonName}{Common name for species}
\item{ScientificName}{Scientific name for species}
\item{ITIS}{Species ITIS Taxon Serial Number }
\item{StockArea}{Stock Area: The region in which the species is assessed}
\item{Year}{Year the data was collected/generated}
\item{Value}{Value of \code{Metric}}
\item{Metric}{Typeof metric: Abundance of species, Index of Biomass, Catch, Recruitment of juvenile individuals, Fmort - Fishing mortality}
\item{Description}{Description of \code{Metric}}
\item{Units}{Units of \code{Value}}
\item{AssessmentYear}{Year of Assessment from which the data was obtained}
\item{FMP}{Fisheries management plan species is managed under}
\item{RegionalEcosystem}{The Regional Ecosystem wher stock is found. There are 8 ecosystems defined around the coastal margins of the United States}
\item{Jurisdiction}{The Fishery Management Council or other group responsible for managing this stock}
\item{UpdateType}{Describes the level of assessment effort.
Used for assessments completed in 2018 and earlier.
New, Benchmark, Full Update, Partial Update}

}
}
\source{
\url{https://www.st.nmfs.noaa.gov/stocksmart?app=browse_by_stock}
}
\usage{
stockAssessmentData
}
\description{
Stock assessment data for species found in waters of the USA. Data is provided for actively managed stocks of a federal Fisheries Management Plan (FMP). Stocks solely managed by US States are not provided.
}
\section{Data download}{


The data were downloaded from \href{https://www.st.nmfs.noaa.gov/stocksmart?app=browse_by_stock}{Stock SMART} - Status, Management, Assessments & Resource Trends. Follow these steps:
\enumerate{
\item Click the "Download Data" module
\item Select data to download (dropdown) -> Assessment Time Series Data
\item Click Search Stocks button
\item Click Select All Stocks button
\item Click Download All button
\item Download all Parts
}
}

\seealso{
Other stockAssessments: 
\code{\link{stockAssessmentSummary}}
}
\concept{stockAssessments}
\keyword{datasets}
