% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stockAssessmentSummary.R
\docType{data}
\name{stockAssessmentSummary}
\alias{stockAssessmentSummary}
\title{Reference points, contact info, assessment metadata for USA assessed stocks}
\format{
A data frame with n rows and m variables

\describe{
\item{Stock Name}{}
\item{Jurisdiction}{}
\item{FMP}{}
\item{Science Center}{}
\item{Regional Ecosystem}{}
\item{FSSI Stock?}{}
\item{ITIS Taxon Serial Number}{}
\item{Scientific Name}{}
\item{Common Name}{}
\item{Stock Area}{}
\item{Assessment Year}{}
\item{Assessment Month}{}
\item{Last Data Year }{}
\item{Update Type}{}
\item{Review Result }{}
\item{Assessment Model}{}
\item{Model Version}{}
\item{Lead Lab }{}
\item{Citation}{}
\item{Final Assessment Report 1}{}
\item{Final Assessment Report 2}{}
\item{Point of Contact}{}
\item{Life History Data}{}
\item{Abundance Data}{}
\item{Catch Data}{}
\item{Assessment Level}{}
\item{Assessment Frequency}{}
\item{Assessment Type}{}
\item{Model Category }{}
\item{Catch Input Data}{}
\item{Abundance Input Data}{}
\item{Biological Input Data}{}
\item{Ecosystem Linkage}{}
\item{Composition Input Data}{}
\item{F Year}{}
\item{Estimated F}{}
\item{F Unit}{}
\item{F Basis}{}
\item{Flimit}{}
\item{Flimit Basis}{}
\item{Fmsy}{}
\item{Fmsy Basis}{}
\item{F/Flimit}{}
\item{F/Fmsy}{}
\item{Ftarget}{}
\item{Ftarget Basis}{}
\item{F/Ftarget}{}
\item{B Year}{}
\item{Estimated B }{}
\item{B Unit }{}
\item{B Basis }{}
\item{Blimit }{}
\item{Blimit Basis}{}
\item{Bmsy}{}
\item{Bmsy Basis}{}
\item{B/Blimit}{}
\item{B/Bmsy}{}
\item{MSY}{}
\item{MSY Unit}{}

}
}
\source{
\url{https://www.st.nmfs.noaa.gov/stocksmart?app=browse_by_stock}
}
\usage{
stockAssessmentSummary
}
\description{
Stock assessment metadata and reference points for species found in waters of the USA.
Data is provided for actively managed stocks of a federal Fisheries Management Plan (FMP). Stocks solely managed by US States are not provided.
For a description of all variables please refer to the Stock SMART data dictionary \url{https://www.st.nmfs.noaa.gov/stocksmart/StockSMART_DataDictionary.pdf}
}
\section{Data download}{


The data were downloaded from {Stock SMART}(https://www.st.nmfs.noaa.gov/stocksmart?app=browse_by_stock) - Status, Management, Assessments & Resource Trends. Follow these steps:
\enumerate{
\item Click the "Download Data" module
\item Select data to download (dropdown) -> Assessment Summary Data
\item Select full range of years from Start Calendar Year- End Calendar Year
\item Click Search Stocks button
\item Click Select All Stocks button
\item Click Select All Fields button
\item Click Download button
\item Download all Parts
}
}

\seealso{
Other stockAssessments: 
\code{\link{stockAssessmentData}}
}
\concept{stockAssessments}
\keyword{datasets}
