% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocksmart-package.R
\docType{package}
\name{stocksmart-package}
\alias{stocksmart}
\alias{stocksmart-package}
\title{stocksmart: Provides access to NOAAs stock SMART data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides access to data found on the NOAA Stock SMART website. The package serves as a data repository for fisheries assessment information and includes utilities for data retrieval.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NOAA-EDAB/stocksmart}
  \item \url{https://noaa-edab.github.io/stocksmart/}
  \item Report bugs at \url{https://github.com/NOAA-EDAB/stocksmart/issues}
}

}
\author{
\strong{Maintainer}: Andy Beet \email{andrew.beet@noaa.gov} (\href{https://orcid.org/0000-0001-8270-7090}{ORCID})

}
\keyword{internal}
